Update to include private vlans, various small enhancements, separation into OrcaNode classes

+++ orca-external/flukes/trunk/src/main/java/orca/flukes/DeleteEdgeMenuItem.java
@@ -16,6 +16,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.JMenuItem;
 
+import com.hyperrealm.kiwi.ui.dialog.KQuestionDialog;
+
 /**
  * A class to implement the deletion of an edge from within a 
  * PopupVertexEdgeMenuMousePlugin.
@@ -32,10 +34,16 @@ public class DeleteEdgeMenuItem<V, E> extends JMenuItem implements EdgeMenuListe
     	ic = i;
         this.addActionListener(new ActionListener(){
             public void actionPerformed(ActionEvent e) {
-            	ic.deleteEdgeCallBack(edge);
-                visComp.getPickedEdgeState().pick(edge, false);
-                visComp.getGraphLayout().getGraph().removeEdge(edge);
-                visComp.repaint();
+            	KQuestionDialog kqd = new KQuestionDialog(GUI.getInstance().getFrame(), "Exit", true);
+        		kqd.setMessage("Are you sure you want to delete edge " + edge.toString() + "?");
+        		kqd.setLocationRelativeTo(GUI.getInstance().getFrame());
+        		kqd.setVisible(true);
+        		if (kqd.getStatus()) {
+        			ic.deleteEdgeCallBack(edge);
+        			visComp.getPickedEdgeState().pick(edge, false);
+        			visComp.getGraphLayout().getGraph().removeEdge(edge);
+        			visComp.repaint();
+        		}
             }
         });
     }
